+++ app/models/category.rb
@@ -1,6 +1,5 @@
 class Category < ActiveHash::Base
   field :accept_to_update, default: true
-  field :segments, default: { short: "SHORT PROGRAM", free: "FREE SKATING"}
   self.data =
     [
      ### senior
@@ -22,7 +21,6 @@ class Category < ActiveHash::Base
      {
        name: "ICE DANCE",
        isu_bio_url: "http://www.isuresults.com/bios/fsbiosicedancing.htm",
-       segments: { short: "SHORT DANCE", free: "FREE DANCE" },
        senior: true,
      },
      #### junior
@@ -31,23 +29,40 @@ class Category < ActiveHash::Base
      { name: "JUNIOR PAIRS"},
      { name: "JUNIOR ICE DANCE"},
     ]
+
+  def update_skaters
+    parser = Parser::SkaterParser.new
+    ActiveRecord::Base.transaction do
+      parser.parse_skaters(name, isu_bio_url).each do |hash|
+        Skater.find_or_create_by(isu_number: hash[:isu_number]) do |skater|
+          skater.update!(hash)
+        end
+      end
+    end  # transaction
+  end
+    
   class << self
-    ## scope
+    ## like scope
     def senior
       #self.all.select {|cat| cat.senior }
       self.where(senior: true)
     end
+    ## class methods
     def senior_categories
       self.senior.map(&:name)
     end
 
-    ## class methods
-    def accept_to_update(categories)
-      Category.all.map {|c| c.accept_to_update = false }
-      [categories].flatten.each do |category|
-        item = find_by(name: category.to_s) || next
-        item.accept_to_update = true
-      end
+    def accept!(categories)
+      all.map {|c| c.accept_to_update = false }   # disable all once and..
+      categories = [categories].flatten.map(&:to_s)
+      where(name: categories).map {|item| item.accept_to_update = true }
     end
-  end
+      
+    def accept?(category)
+      where(accept_to_update: true, name: category.to_s).present?
+    end
+    def update_skaters
+      all.select(&:isu_bio_url).map(&:update_skaters)
+    end
+  end ## self
 end